drm: rockchip: inno_hdmi: clear the hpd interrupt directly in irq function
Somehow if we return the IRQ_WAKEUP_THREAD flag in irqhandler function,
the irq thread just wouldn't wakeup. Due to driver only clear the HPD
interrupt in irq thread, so there would be an endless HPD interrupt.

So let's just clear the HPD interrupt in irqhandler, and once the irq
thread wakeup, then we just report a hotplug event to drm directly.

Signed-off-by: Yakir Yang <ykk@rock-chips.com>