FROMLIST: drm/rockchip: Send events for same-fb flips
Rockchip previously treated a pageflip to the same framebuffer as a
no-op, discarding the event if one was requested. This breaks Weston,
which, when idle, sends a no-op vblank event to discover vblank
timings if the vblank query interface is not usable.

Silently dropping events is also quite a hostile thing to do to
userspace in general.


Signed-off-by: Daniel Stone <daniels@collabora.com>
Cc: Sjoerd Simons <sjoerd.simons@collabora.co.uk>
Cc: Heiko Stuebner <heiko@sntech.de>
Tested-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>
Tested-by: Heiko Stuebner <heiko@sntech.de>
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
(am from https://patchwork.kernel.org/patch/7624901/)