FROMLIST: usb: dwc2: host: Keep track of and use our scheduled microframe
The microframe scheduler did a lot of work to pick the proper
microframe.  Then dwc2_sched_periodic_split() went ahead and ignored
which microframe we picked if it ever needed to re-assign things.

Let's keep track of the uframe and then we'll always use it.


Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
(am from https://patchwork.kernel.org/patch/7631561/)