ARM: dts: rk3036: override operating-points to 600M on kylin board
That's a hack way to make the emac working.
Let the CPU can divide out the clock since the emac depends on
the 50M clock to work.

Moment, we add the hacy way to make the emac working firstly.
We can revert it if we find the apll or other issues.

Signed-off-by: Caesar Wang <wxt@rock-chips.com>