UPSTREAM: clk: rockchip: fix section mismatches with new child-clocks
To model the muxes downstream of fractional dividers we introduced the
child property, allowing to describe a direct child clock.
The first implementation seems to cause section warnings, as the core
clock-tree is marked as initdata while the data pointed to from the
child element is not.

While there may be some way to also set that missing property in the
inline notation I didn't find it, so to actually fix the issue for now
move the sub-definitions into separate declarations that can have
their own __initdata properties.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Signed-off-by: Michael Turquette <mturquette@baylibre.com>

(cherry picked from git.kernel.org mmind/linux-rockchip.git v4.6-armsoc/dts32
 commit 5b73840375e3eebeb7adf957ff64a96abdf4e1a1)
Signed-off-by: Caesar Wang <wxt@rock-chips.com>