FROMLIST: ARM: dts: rockchip: add the sdmmc for kylin board
We should add the sdmmc but enabling it, since the sdmmc and
uart2 ie reusing the same pin.
That will cause the debug information can't display on console.

Anyway, you only need add the okay status if someone wants to enable
the sdmmc.

e.g.
if you use the adb to debug and run with android os.
you can add the
status = "okay" to enable the sdmmc for sd card working.

Change-Id: Ic8163e981c2c60a7d6f9871aae7ab9322efcc5d5
Signed-off-by: Caesar Wang <wxt@rock-chips.com>