FROMLIST: clk: rockchip: rk3036: Add to 100MHz integer multiples of cpu freq table
We need the cpu changing freqs capability, and considering that emac ref
which is need 50MHz might is under the APLL, meanwhile, we don't need
too many freqs because it may spend more time to find the correct freq.
Therefore, we should add 100MHz integer multiples cpu freqs for emac.

Change-Id: I53b12ef0c211f0d8d59ec572056a391836a67072
Signed-off-by: Xing Zheng <zhengxing@rock-chips.com>