UPSTREAM: drm/rockchip: vop: fix mask when updating interrupts
Commit dbb3d94 (drm/rockchip: vop: move interrupt registers into
vop_data) introduced new macros for updating the interrupt control
registers but these always use the mask from the register definition
without refining it for the particular bits that are being changed.

This means that whenever we enable/disable a particular interrupt we end
up disabling all of the others as a side effect.

Signed-off-by: John Keeping <john@metanate.com>

(cherry picked from github markyzq/kernel-drm-rockchip drm-rockchip-next-2016-02-18
 commit c7647f8681feeb6c0957e3cf5daed1fbf8b3a5af)
Signed-off-by: Caesar Wang <wxt@rock-chips.com>